Remove transitive dependency in ActionWrapper
