+++ backend/src/database/data-source.ts
@@ -1,8 +1,8 @@
 import { DataSource } from 'typeorm';
-import { ActionEmitter } from '../exercise/action-emitter';
-// import { ActionWrapper } from '../exercise/action-wrapper';
+// import { ActionEmitter } from '../exercise/action-emitter';
+import { ActionWrapper } from '../exercise/action-wrapper';
 import { Config } from '../config';
-import { ActionWrapper, ExerciseWrapper } from '../exercise/exercise-wrapper';
+import { ActionEmitter, ExerciseWrapper } from '../exercise/exercise-wrapper';
 
 export const createNewDataSource = () => {
     Config.initialize();
