+++ backend/src/database/services/action-emitter.service.ts
@@ -1,19 +1,35 @@
-import type { EntityManager } from 'typeorm';
+import type { DataSource, EntityManager } from 'typeorm';
+import { ActionEmitter } from '../../exercise/exercise-wrapper';
 import type { Creatable, Updatable } from '../dtos';
-import { ActionEmitter } from '../../exercise/action-emitter';
+// import { ActionEmitter } from '../../exercise/action-emitter';
 import { BaseService } from './base-service';
+import type { ExerciseWrapperService } from './exercise-wrapper.service';
 
 export class ActionEmitterService extends BaseService<ActionEmitter> {
-    protected createSavableObject<TInitial extends ActionEmitter | undefined>(
+    public constructor(
+        private readonly exerciseService: ExerciseWrapperService,
+        dataSource: DataSource
+    ) {
+        super(dataSource);
+    }
+    protected async createSavableObject<
+        TInitial extends ActionEmitter | undefined
+    >(
         initialObject: TInitial,
         dto: TInitial extends ActionEmitter
             ? Updatable<ActionEmitter>
             : Creatable<ActionEmitter>,
         manager: EntityManager
-    ): ActionEmitter {
-        return initialObject
-            ? manager.merge(this.entityTarget, initialObject, dto)
-            : manager.create(this.entityTarget, dto);
+    ): Promise<ActionEmitter> {
+        const { exerciseId, ...rest } = dto;
+        const emitter = initialObject
+            ? manager.merge(this.entityTarget, initialObject, rest)
+            : manager.create(this.entityTarget, rest);
+        emitter.exercise =
+            initialObject && exerciseId === undefined
+                ? emitter.exercise
+                : await this.exerciseService.findById(exerciseId!, manager);
+        return emitter;
     }
 
     protected readonly entityTarget = ActionEmitter;
