+++ backend/src/database/services/action-wrapper.service.ts
@@ -1,17 +1,15 @@
 import type { EntityManager, DataSource } from 'typeorm';
 import type { UUID } from 'digital-fuesim-manv-shared';
 import type { Creatable, Updatable } from '../dtos';
-// import { ActionWrapper } from '../../exercise/action-wrapper';
-import type { ActionEmitter } from '../../exercise/action-emitter';
-import { ActionWrapper } from '../../exercise/exercise-wrapper';
+import { ActionWrapper } from '../../exercise/action-wrapper';
+// import type { ActionEmitter } from '../../exercise/action-emitter';
+import type { ActionEmitter } from '../../exercise/exercise-wrapper';
 import type { ActionEmitterService } from './action-emitter.service';
 import { BaseService } from './base-service';
-import type { ExerciseWrapperService } from './exercise-wrapper.service';
 
 export class ActionWrapperService extends BaseService<ActionWrapper> {
     public constructor(
         private readonly actionEmitterService: ActionEmitterService,
-        private readonly exerciseWrapperService: ExerciseWrapperService,
         dataSource: DataSource
     ) {
         super(dataSource);
@@ -26,7 +24,7 @@ export class ActionWrapperService extends BaseService<ActionWrapper> {
             : Creatable<ActionWrapper>,
         manager: EntityManager
     ): Promise<ActionWrapper> {
-        const { emitter: actionEmitter, exerciseId, ...rest } = dto;
+        const { emitter: actionEmitter, ...rest } = dto;
         const actionWrapper = initialObject
             ? manager.merge(this.entityTarget, initialObject, rest)
             : manager.create(this.entityTarget, rest);
@@ -48,9 +46,6 @@ export class ActionWrapperService extends BaseService<ActionWrapper> {
                   actionEmitter as Creatable<ActionEmitter>,
                   manager
               );
-        actionWrapper.exercise = exerciseId
-            ? await this.exerciseWrapperService.findById(exerciseId, manager)
-            : actionWrapper.exercise;
         return actionWrapper;
     }
